Updating the delete function

+++ app.py
@@ -89,10 +89,11 @@ def update_user(user_id):
 @login_required
 def delete_user(user_id):
     user = User.query.get(user_id)
+    if user_id == current_user.id:
+        return jsonify({'message': 'Deleção não permitida'}), 403
     if user:
         db.session.delete(user)
         db.session.commit()
-        logout()
         return jsonify({'message': f'User {user_id} deleted successfully'})
     return jsonify({'message': 'User not found'}), 404
